moved Cache and ChainCache structu into caching.rs
