+++ cli/src/cmd/forge/cache.rs
@@ -4,9 +4,10 @@ use clap::{Parser, Subcommand};
 use std::str::FromStr;
 
 use crate::cmd::Cmd;
+use caching::{Cache, ChainCache};
 use ethers::prelude::Chain;
 use eyre::Result;
-use foundry_config::{Cache, Chain as FoundryConfigChain, ChainCache, Config};
+use foundry_config::{caching, Chain as FoundryConfigChain, Config};
 
 #[derive(Debug, Parser)]
 pub struct CacheArgs {
