+++ config/src/lib.rs
@@ -21,13 +21,11 @@ use figment::{
     Error, Figment, Metadata, Profile, Provider,
 };
 use inflector::Inflector;
-use number_prefix::NumberPrefix;
 use semver::Version;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::{
     borrow::Cow,
     collections::{hash_map::Entry, BTreeSet, HashMap},
-    fmt::Formatter,
     fs,
     path::{Path, PathBuf},
     str::FromStr,
@@ -41,6 +39,7 @@ pub mod utils;
 pub use crate::utils::*;
 
 pub mod caching;
+use caching::{Cache, ChainCache};
 mod chain;
 pub use chain::Chain;
 
@@ -949,49 +948,6 @@ impl Config {
         }
         eyre::bail!("Could not read the path, it is probably not a directory")
     }
-}
-
-/// Content of the foundry cache folder
-#[derive(Debug)]
-pub struct Cache {
-    /// The list of chains in the cache
-    pub chains: Vec<ChainCache>,
-}
-
-impl std::fmt::Display for Cache {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        for chain in &self.chains {
-            match NumberPrefix::decimal(chain.blocks.iter().map(|x| x.1).sum::<u64>() as f32) {
-                NumberPrefix::Standalone(size) => {
-                    writeln!(f, "-️ {} ({:.1} B)", chain.name, size)?;
-                }
-                NumberPrefix::Prefixed(prefix, size) => {
-                    writeln!(f, "-️ {} ({:.1} {}B)", chain.name, size, prefix)?;
-                }
-            }
-            for block in &chain.blocks {
-                match NumberPrefix::decimal(block.1 as f32) {
-                    NumberPrefix::Standalone(size) => {
-                        writeln!(f, "\t-️ Block {} ({:.1} B)", block.0, size)?;
-                    }
-                    NumberPrefix::Prefixed(prefix, size) => {
-                        writeln!(f, "\t-️ Block {} ({:.1} {}B)", block.0, size, prefix)?;
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
-/// A chain folder in the foundry cache
-#[derive(Debug)]
-pub struct ChainCache {
-    /// The name of the chain
-    pub name: String,
-
-    /// A tuple containing block number and the block directory size in bytes
-    pub blocks: Vec<(String, u64)>,
 }
 
 impl From<Config> for Figment {
